Moved space and neighbor initialization to init
If we want to extend the environment by changing the observation space/action space/neighbor configuratin, this information will be reset when we reset the environment. 

Also initialize the open_cells, mines, steps and unneccesary_steps. Normally when creating an environment the first thing you do is call reset, but this will ensure that the game can be run properly when reset is not called after __init__